Removed majority of deprecated functions in light of newer faster PHP5 functions.
git-svn-id: http://repos.svn.fusionticket.org/fusionticket/trunk@545 87ea6b0e-88cb-4cdd-95e9-9c88242927a5

+++ includes/admin/AdminView.php
@@ -408,15 +408,17 @@ class AdminView extends AUIComponent {
             <td class='admin_value'>";
 
             if ($type == 'img') {
-                 list($width, $height, $type, $attr) = getimagesize(ROOT.'files'.DS.$data[$name]);
-
-                 if (($width>$height) and ($width > 300)) {
-                   $attr = "width='300'";
-                 } elseif ($height > 250) {
-                   $attr = "height='250'";
-                 }
-
-                echo "<img $attr src='$src'>";
+           		if(file_exists(ROOT.'files'.DS.$data[$name])){
+           			list($width, $height, $type, $attr) = getimagesize(ROOT.'files'.DS.$data[$name]);
+					if (($width>$height) and ($width > 300)) {
+						$attr = "width='300'";
+					} elseif ($height > 250) {
+						$attr = "height='250'";
+					}
+					echo "<img $attr src='$src'>";	
+           		}else{
+           			echo "<strong>File does not exsist</strong>";
+           		}
             } else {
                 echo "<a href='$src'>{$data[$name]}</a>";
             }
